Add notes for a walkthrough of a Sail-x86 demo.
I've tested this on a FreeBSD laptop (that I will bring with me), but
this should also generally work on macOS (I've added some notes about
running on macOS).

Note that this does not assume running on Morello as CheriBSD package
repositories don't have the packages available (even as hybrid).